rewrote associated hero function
