+++ test/test.js
@@ -548,7 +548,7 @@ describe("web", function() {
                 done(err);
             });
         });
-        it('should return JSON result', function(done) {
+        it('should return JSON', function(done) {
             request.get(process.env.ROOT_URL + '/api/matches?draw=2&columns%5B0%5D%5Bdata%5D=match_id&columns%5B0%5D%5Bname%5D=&columns%5B0%5D%5Bsearchable%5D=true&columns%5B0%5D%5Borderable%5D=true&columns%5B0%5D%5Bsearch%5D%5Bvalue%5D=&columns%5B0%5D%5Bsearch%5D%5Bregex%5D=false&columns%5B1%5D%5Bdata%5D=game_mode&columns%5B1%5D%5Bname%5D=&columns%5B1%5D%5Bsearchable%5D=true&columns%5B1%5D%5Borderable%5D=true&columns%5B1%5D%5Bsearch%5D%5Bvalue%5D=&columns%5B1%5D%5Bsearch%5D%5Bregex%5D=false&columns%5B2%5D%5Bdata%5D=cluster&columns%5B2%5D%5Bname%5D=&columns%5B2%5D%5Bsearchable%5D=true&columns%5B2%5D%5Borderable%5D=true&columns%5B2%5D%5Bsearch%5D%5Bvalue%5D=&columns%5B2%5D%5Bsearch%5D%5Bregex%5D=false&columns%5B3%5D%5Bdata%5D=duration&columns%5B3%5D%5Bname%5D=&columns%5B3%5D%5Bsearchable%5D=true&columns%5B3%5D%5Borderable%5D=true&columns%5B3%5D%5Bsearch%5D%5Bvalue%5D=&columns%5B3%5D%5Bsearch%5D%5Bregex%5D=false&columns%5B4%5D%5Bdata%5D=start_time&columns%5B4%5D%5Bname%5D=&columns%5B4%5D%5Bsearchable%5D=true&columns%5B4%5D%5Borderable%5D=true&columns%5B4%5D%5Bsearch%5D%5Bvalue%5D=&columns%5B4%5D%5Bsearch%5D%5Bregex%5D=false&columns%5B5%5D%5Bdata%5D=parse_status&columns%5B5%5D%5Bname%5D=&columns%5B5%5D%5Bsearchable%5D=true&columns%5B5%5D%5Borderable%5D=true&columns%5B5%5D%5Bsearch%5D%5Bvalue%5D=&columns%5B5%5D%5Bsearch%5D%5Bregex%5D=false&order%5B0%5D%5Bcolumn%5D=0&order%5B0%5D%5Bdir%5D=asc&start=0&length=10&search%5Bvalue%5D=&search%5Bregex%5D=false&_=1422621884994', function(err, resp, body) {
                 assert(resp.statusCode === 200);
                 assert(body);
@@ -573,30 +573,31 @@ describe("web", function() {
         });
     });
     describe("/preferences", function() {
-        it('should load', function(done) {
+        it('should return JSON', function(done) {
             request.post(process.env.ROOT_URL + '/preferences', {}, function(err, resp, body) {
-                //todo assert
+                assert(body)
                 done(err);
             });
         });
     });
     describe("/fullhistory", function() {
-        it('should load', function(done) {
+        it('should return JSON', function(done) {
             request.post(process.env.ROOT_URL + '/fullhistory', {}, function(err, resp, body) {
-                //todo assert
+                assert(body);
                 done(err);
             });
         });
     });
-    describe("/verify_captcha", function() {
-        it('should load', function(done) {
-            request.post(process.env.ROOT_URL + '/verify_captcha', {
+    describe("/verify_recaptcha", function() {
+        it('should return JSON', function(done) {
+            request.post(process.env.ROOT_URL + '/verify_recaptcha', {
                 form: {
-                    recaptcha_challenge_field: "",
-                    recaptcha_response_field: ""
+                    recaptcha_challenge_field: "asdf",
+                    recaptcha_response_field: "jkl;"
                 }
             }, function(err, resp, body) {
-                //todo assert
+                assert(resp.statusCode === 200);
+                assert(body);
                 done(err);
             });
         });
@@ -617,12 +618,6 @@ describe("tasks", function() {
             done(err);
         });
     });
-    it('untrack players', function(done) {
-        tasks.untrackPlayers(function(err, num) {
-            assert.equal(num, 1);
-            done(err);
-        });
-    });
     it('generate constants', function(done) {
         tasks.generateConstants(function(err) {
             done(err);
@@ -632,16 +627,6 @@ describe("tasks", function() {
         tasks.getFullMatchHistory(function(err) {
             done(err);
         }, ["1"]);
-    });
-    it('clear active jobs', function(done) {
-        jobs.inactive(function(err, ids) {
-            kue.Job.get(ids[0], function(err, job) {
-                job.active();
-                tasks.clearActiveJobs(function(err) {
-                    done(err);
-                });
-            })
-        })
     });
 });
