+++ views/player_index.jade
@@ -22,6 +22,7 @@ block playercontent
            th Match ID
            th Hero
            th Result
+           th Region
            th Game Mode
            th Duration
            th Played
@@ -39,6 +40,7 @@ block playercontent
               else
                 =match.players[0].hero_id
             td= match.player_win ? "Won" : "Lost"
+            td= constants.regions[match.cluster] || match.cluster
             td= constants.modes[match.game_mode] ? constants.modes[match.game_mode].name : match.game_mode
             td= moment().startOf('day').seconds(match.duration).format("H:mm:ss")
             td= moment.unix(match.start_time + match.duration).fromNow()
